srcs/requirements/backend/django/requirements.txt
Django==4.2
djangorestframework
psycopg2
djangorestframework-simplejwt
